Restrict write access to vendor/*
You shouldn't be editing braided things, and you should be thinking carefully
about commits you make to the rest of the code in vendor. This will stop the
commit and make you run with --no-verify if you've touched something risky.

Lucky for us, Braid doesn't run pre-commit hooks, so braid update will continue
to work as normally as it ever has.