First version of prepare-commit-msg hook
At this stage just handles the simple case of staging changes to a
submodule and including the list of changes in the standard commit
template. Right now, we only consider changes staged in the index
relative to the current HEAD; in a subsequent commit we'll make it
handle other points of comparison (as this will come up when committing
with --amend or -c/-C).

This will evidently need to be tweaked as we discover how it interacts
with the different kinds of commit message source.