Don't show summaries for merge commits
The only sane way to report submodule changes for merge commits would be
to diff against the merge base, but that may be many commits back for
long-lived topic branches, and could produce unmanageably large commit
messages.

So, we just don't record summaries at all for merge commits. This is ok
because the parent commits that lead up to the merge commit should have
recorded all the submodule changes along the way anyway.