Add commit-msg hook from Gerrit
This will add a unique Change-Id: to your commit messages so that Gerrit
can recognize revisions of already-submitted patches (ie. the first time
you submit your commit will receive a unique id; if you later amend and
then resubmit your commit will retain that same identifier).

This is preparatory for our forthcoming experiment with Gerrit. We'll
first be trialling Gerrit with a small app (ie. not causes.git). If we
like it, we can move to it for other apps too. If not, this hook and the
Change-Id: are a harmless modifications (ie. the release notes parser in
Starways already knows to strip out Change-Ids when preparing release
notes).