Add ERB syntax check to pre-commit hook
Just noticed a syntax error in an ERB file had slipped past us into
production:

  https://github.com/causes/causes/commit/ba5bc23c91dce9813f976a5165292327e0b83faf

Sounds like a job for git_hooks.

Note that in order to check ERB files for syntax errors, we have to
"compile" them first, which means that the line numbers reported by the
hook probably won't be very helpful (they'll be the line numbers in the
compiled source, not the ERB template itself), but at least you'll get a
heads up, a hint, and also the name of the template that contains the
error.

Change-Id: I95400449dd1c69ceb17993ac281a89636aa430e6