Always strip submodule changes before reinserting them
Previously, we tried to be conservative about when we removed submodule
changes from a commit message.

I've now seen quite a few duplicates sneak through, which are probably
the result of rebasing or amending, and the submodule changes being
redundantly added again.

As the regex which detects the changes is pretty tight, I don't think
there's any real risk of false positives, so we can just unconditionally
strip submodule changes before we ever insert them.

Change-Id: I1d138c0dca7c1b5d040970f3d52f4cba3f10c2da
Reviewed-on: https://gerrit.causes.com/570
Tested-by: Greg Hurrell <greg@causes.com>
Reviewed-by: Jimmy Kittiyachavalit <jimmy@causes.com>