Remove commented-out code from post-merge hook
Automatically running migrations is not a good idea as it is a
potentially destructive operation, adds a lengthy delay to the
post-merge hook, and is inapplicable to non-Rails projects anyway.

The "right" way to do this would be to run "rake
db:abort_if_pending_migrations" and prompt the user to run the
migrations manually, but on an app like Causes running any Rake task
involves a lengthy boot-up of Rails, which would introduce an
unacceptable delay.

Change-Id: I6bc7a048797b4516496f0da2f311b90a57c4d454
Reviewed-on: https://gerrit.causes.com/939
Tested-by: Greg Hurrell <greg@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>