Add test verification to pre-commit hook
Our current pre-commit hook simply shows the user the list of tests they
SHOULD have run. It does not do any sort of verification on whether or
not these tests actually were run.

This commit adds support to the pre-commit hook--along with the use of
the special spec formatter which records which tests were run--to verify
that tests actually were run by comparing the timestamps of the last
test run to the modification times of the files that were modified.

Release Note: (Eng) The pre-commit hook will now verify you have run
tests.
  http://www.pivotaltracker.com/story/show/21326941
  ("Add test verification to pre-commit hook")

Change-Id: I9c05c3395279271777c587cf08c0493a6bec7f93
Reviewed-on: https://gerrit.causes.com/1327
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>