Teach git_hooks to reject console.log
Calls to console.log lead to Javascript errors in Internet Explorer
unless Developer Tools is turned on.

With this check in place, bad commits will see feedback like this:

  Checking js_console_log...................FAILED
    example.js:2:    console.log("random goop!");

Note that this is just a dumb grep for the string "console.log", and
doesn't try to do anything clever like excluding commented-lines. That
should be fine, as nobody should be committing commented-out code
anyway.

Related story:
  http://www.pivotaltracker.com/story/show/18119495
  ("Teach git_hooks to reject console.log")

Change-Id: Id3a97de84a4f65100e7281e6ce9bbca3c900fcb3
Reviewed-on: https://gerrit.causes.com/1585
Tested-by: Greg Hurrell <greg@causes.com>
Reviewed-by: Kristján Pétursson <kristjan@causes.com>
Reviewed-by: Chris Chan <chris@causes.com>