Add call to strip for each line in the .spec-results file
Apparently our earlier assumption that a call to "strip" was not
necessary was invalid. If you don't call strip, each file path
string will have a newline character appended to it, resulting in
later set operations not working as expected since the newline
makes the strings inequal.

So we include the call to strip, and throw it in its own block since
it doesn't seem to work otherwise (as in an error occurs due to the
difference in the version of ruby running, since the hooks aren't
executed using bundler).

Change-Id: Ib403463b05d14885294d500ec7627a2d0b7ca41b
Reviewed-on: https://gerrit.causes.com/1616
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>