Fix install script for hooks which do not yet exist
Previously, running the install script would fail if a repo did not
already have a script for a particular commit hook in its `.git/hooks`
directory.

This was caused by commit 750b64e, which added a call to `FileUtils.rm`,
which fails if a file does not exist. The fix was to change this to
`FileUtils.rm_f`, so that it does not fail if the file does not exist.

Change-Id: I16ed4418962824b48c0772b34a7bc8ba0558d390
Reviewed-on: https://gerrit.causes.com/12276
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Kim Douglas Clarence Mason <kim.mason@causes.com>