Skip ctags on stock OS X
The `ctags` binary provided by Apple isn't sufficient, and doesn't
support any of the arguments we try to send it. It also doesn't respond
to --version, making it easy to fail fast.

Will add to documentation that this requires installing a 'normal'
ctags, e.g. via homebrew, for tagging to work on OS X.

Change-Id: Icd16919458c1734f95b28c17468282e9a477e2f7
Reviewed-on: https://gerrit.causes.com/22354
Tested-by: Aiden Scandella <aiden@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>