Add Configuration with templates
Part of the default behavior of this project will change when it is
released as a gem, since we no longer wish to install *all* of the hooks
in here by default.

I create a 'default' template, which ignores a subset of checks that are
Causes/gerrit-specific.

Not used yet, but soon...

Change-Id: I807601c2b8a83477ad72eeff198ae218c780d4fb
Reviewed-on: https://gerrit.causes.com/22692
Tested-by: Aiden Scandella <aiden@causes.com>
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>