Add un-skippable checks
The environment variable SKIP_CHECKS is convenient to skip pre-commit
hooks when the developer knows what he or she is doing. The special
value of 'all' is used to signify that all checks should be skipped, but
we really want it to mean 'all of the safety checks' should be skipped.

This commit adds a class-level annotation to HookSpecificCheck and
changes the way we load plugins ever so slightly, so that the class can
be introspected before we decide we really want to skip it.

Fixes #1.

Change-Id: I85f05f35f093909ee729b823de68b3a990b92d66
Reviewed-on: https://gerrit.causes.com/22974
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>