Fix Reporter#with_status width
This got accidentally a while ago when the lint names became modularized
(`Overcommit::GitHook::HookName`), because name includes the modules.
This uses the `friendly_name` which is the same as what will be output
to the console.

At the same time, I reduced the max width from 70 to 60, since 70 seemed
too wide.

Change-Id: Idace958b82918ad6b67ae294bd36b6e7bc272bfc
Reviewed-on: https://gerrit.causes.com/23515
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Lann Martin <lann@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>