Fix bug in js_syntax where `output` should be string
After resurrecting StagedFile in 0.1.8, `output` is passed as an
argument to `filter_string` which expects a string. However, since it
was being split on newlines on the previous line, it was an array. This
caused overcommit to explode like:

  `filter_string': undefined method `gsub' for []:Array (NoMethodError)

To fix this, we removed the `split` and changed a call to `.none?` to
`.empty?`.

This bumps the version from 0.1.9 to 0.1.10.

Change-Id: I3552a6dd6cd61c599428fd7e2388d71e57d626e3
Reviewed-on: https://gerrit.causes.com/23994
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>