Teach StagedFile which lines were changed
In order to support complex hook checks such as only returning an error
when a lint is detected on a line that a user touched, a `lines_changed`
method was added to the `StagedFile` class, which returns a set of
numbers representing the lines that were changed (1-based).

In the process, specs where added for the `StagedFile` class (as it
previously didn't have any). In order to support this, a module
`GitSpecHelpers` was added with functionality to make working with `git`
repositories a little easier. This was necessary to test `StagedFile`,
since its implementation is tightly coupled to using `git`.

Change-Id: Id3781a994d670f83979e08d77035fec97d0345ee
Reviewed-on: https://gerrit.causes.com/25158
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>