Teach Ruby style check to treat unmodified lines as warnings
This check can get a bit noisy when working in a file with a lot of
lints which weren't caused by you. Rather than have the check report
errors for all style offenses in the file, we change the behaviour to
only count errors for lines actually touched by the diff.

Change-Id: Iafb49ac3ea4769c9541fd4910dee28e3eba9d506
Reviewed-on: https://gerrit.causes.com/25164
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>