Indent check output lines under check name
When overcommit reported errors or warnings, its output format was a bit
odd since it would only indent the first line of output. This modifies
the output to add an indent after every newline character (except the
last) so that all output appears nested under the name of the check.

I also switched an unrelated string to use single quotes.

Change-Id: I0f2d3522eca3c4d475f965dd6b014a0c388c8f6b
Reviewed-on: https://gerrit.causes.com/25556
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Aaron Neyer <aaron.neyer@causes.com>