Change top-level hook types to use CamelCase
Previously, the hook type names were snake case (`pre_commit`, etc).

This was a bit annoying as it was difficult to remember to switch
between underscores for the group name and camel case for the hook name.

Convert the naming convention to use camel case. Note that individual
settings are still snake case to make them further distinct from the
other camel-cased entries.

Change-Id: I47f7dd83452de43684d9d01b43464ba49f468592
Reviewed-on: http://gerrit.causes.com/35596
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>