Fix warning messages in HookMatcher on Ruby 1.8.7
Ruby 1.8.7 prints out the following warning wing when using a hook
matcher with no arguments:

    warning: multiple values for a block parameter (0 for 1)

In an effort to quiet these warnings, convert the matchers to accept an
argument splat.

This change also sets us up to easily extend the matchers to support
additional options in the future, should we elect to go down that route.

Change-Id: Ie4e973a895bb1ad6084d1de87028addf4d61a886
Reviewed-on: http://gerrit.causes.com/36061
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>