Print notice during inline install of gem
When a user attempts to run an overcommit hook without the `overcommit`
gem being installed, the hook file would ask the user if they wanted to
install the gem. If the user entered "y" and enter, it was possible for
the installation to appear "frozen" as no output would occur for quite
some time while the gems are fetched from Rubygems.

A simple solution is to display an "Installing..." message the moment
the user hits enter, so that they know they should wait (rather than
cancelling assuming something is frozen).

Change-Id: Id660073627258439b6424d6a277f34a4bb88d170
Reviewed-on: http://gerrit.causes.com/36177
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>