Update overcommit-hook before hook run if changed
Previously, if upgrading Overcommit would result in the master hook
being changed, the user would have to manually run `overcommit
--install` in order to update the hooks.

Make this a little bit easier by automatically updating the master hook
if it has changed.

One trick to get this to work is that if the hook is changed and we
update it, we need to `exec` in order to replace the current process
with a new one executing the new master hook, otherwise the old file
will still be run.

Change-Id: I6746e6f25cf624c0da9d91153ba1f91b9a9b8e75
Reviewed-on: http://gerrit.causes.com/37388
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>