Add --force-exclusion flag to Rubocop hook
There was a problem where the Overcommit Rubocop check would not exclude
files excluded in the `.rubocop.yml` file.

This was a bug in Rubocop where any file explicitly passed to it on the
command line (which is what Overcommit does) would still be linted even
if it was excluded by the config.

Rubocop 0.20.0 introduces a `--force-exclusion` flag to address this
problem. Include it when we invoke Rubocop in the hook.

Change-Id: I530a4ae96b3431f9104dfd8eb8569623a975f964
Reviewed-on: http://gerrit.causes.com/37398
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>