Ignore `skip` configuration setting in HookSigner
When a user had `verify_plugin_signatures` enabled and specified that a
hook should be skipped via the `SKIP` environment variable, they would
get an erroneous warning about the configuration having changed.

This was due to the `HookSigner` detecting that the `skip` value used by
overcommit had changed. Prevent this by ignoring that particular key.

Change-Id: I5552c315df4b7114c7c437a6332abe05c398666e
Reviewed-on: http://gerrit.causes.com/37942
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>