Restore merge state in pre-commit HookContext
If a user was in the middle of resolving a merge conflict, Overcommit
would blow away the state of the merge since it uses `git stash` to save
changes.

This had the nasty effect of resulting in the user receiving the
following error after attempting to commit changes resolving a merge
conflict:

    fatal: could not open '.git/MERGE_HEAD' for reading: No such file or
    directory

The solution is to explicitly check if we are in a merge state before
stashing, record the relevant state, and restore it after the stashed
changes have been restored.

An integration test was added since there's no particularly good way to
unit test this behavior.

Change-Id: I973487f9bbb1a02d41204a52cc5e5c2eaa6e3a63
Reviewed-on: http://gerrit.causes.com/38035
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>