Gracefully handle committing to empty repo
Previously, when attempting to use `overcommit` with an empty repo (i.e.
you were trying to commit the "Initial commit"), you would get an error
saying that `HEAD` was not defined.

This is caused by the fact that `HEAD` is only set once the branch has
at least one commit.

Solve by explicitly checking for this case and only stashing when we are
not the initial commit (since the initial commit will always only have
staged files, no "modified" files).

This also modifies the implementation of `any_changes?` in the
`PreCommit` hook context to use `git status` instead of `git diff`,
which prevents it from blowing up in the initial commit situation.

Change-Id: Ia5fe7cbe373938f7c4f3fa7c6ae40641b5c40169
Reviewed-on: http://gerrit.causes.com/40324
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>