Improve robustness of signal handling
The previous implementation was technically still vulnerable to problems
since the signal could be received in between the call to
`@context.setup_environment` and `run_hooks`, which could leave the
system in a weird state.

Solve the problem by creating a utility class `InterruptHandler` which
provides an interface for enabling/disabling signal handling so the
`HookRunner` can have more control over when/how the signal is handled.

Change-Id: I411433169ea1bf23f6f6426e67996d5daf7ae394
Reviewed-on: http://gerrit.causes.com/40326
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>