Enable wildcards in globs to match dotfiles
The default behavior of Ruby's glob implementation is for the asterisk
`*` to not match files beginning with a dot by default (e.g. `.profile`,
etc.)

While this is also the default behavior in shells, it's somewhat
inconvenient since there are many dotfiles in modern projects which are
still types of files under some other extension (e.g. `.rubocop.yml` is
a YAML file).

Change the default behavior of globs in Overcommit to match dotfiles by
default, as this is likely the desired behavior in most cases. Users can
always explicitly exclude dotfiles separately if they want to.

Change-Id: Id613d8b95bfb9a15ad2b9dbb2823e391d9823210
Reviewed-on: http://gerrit.causes.com/40420
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>