Add UTF8 encoding marker in Printer class
Since we moved the strings from `HookRunner` to `Printer`, we also need
to mark that file as UTF8 for older versions of Ruby.

In order to get the build to pass, also limit the RuboCop `Encoding`
check to only warn when needed, so that we don't get lots of warnings
under JRuby.

Change-Id: I1eec4710f73c542fa1959ea2a50b8495e96d273e
Reviewed-on: http://gerrit.causes.com/40462
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>