Drop support for Ruby 1.8.7
The time has finally come. While Ruby 1.8.7 technically EOLed a while
ago, we were continuing to support it so users running older versions of
OS X operating systems (and many Linux distributions) would be able to
install it without doing anything.

This was a pain, but not too much hassle. However, a dependency for our
build infrastructure, RuboCop, depends on a gem that does not support
Ruby 1.8.7. It's not worth it for us to jump through hoops to continue
to support 1.8.7, so let's call it a good run and end it now.

At some point we'll follow up with converting all the 1.8 hashes to 1.9
syntax, and removing any other 1.8.7 hacks.

Change-Id: I9a048062bd905c685b8e5a633ca44af7cd1a7879
Reviewed-on: http://gerrit.causes.com/40470
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>