Add ability to change hook statuses
Add ability to change the status of a hook via the `on_fail` and
`on_warn` hook configuration options.

This is intended to give users a greater level of control over the
severity of hooks, since there may be cases where they might want to
report warnings as errors, or vice versa.

Change-Id: Ib59c2d880edc109068717800e7a35a61335894e3
Reviewed-on: http://gerrit.causes.com/40977
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>