Add color to --list-hooks output
A little color makes it easier to see which hooks are enabled/disabled.

While here I refactored the implementation to avoid the use of a
blacklist of keys to ignore--this usually becomes a pain to maintain as
every time you add a key that must be ignored you must remember to
include it in the blacklist. Instead, we use the already existing
`supported_hook_type_class` method from the `Utils` module, which
accomplishes the same thing.

I also removed the need to instantiate a `HookContext` class in order to
determine whether or not the hook was enabled.

Change-Id: Idca7b8e682984b0d9b37b73e912da0ce746bc017
Reviewed-on: http://gerrit.causes.com/42261
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>
Tested-by: Shane da Silva <shane.dasilva@brigade.com>