Add BrokenSymlinks pre-commit hook
The recent bug in the pre-commit `HookContext`'s handling of broken
symlinks prompted the need for a hook that detects these problems
explicitly (otherwise the user might get a cryptic error in one of the
other hooks).

Being explicitly told that a symlink is broken is useful, so add it as a
pre-commit hook.

Change-Id: I1a1330bdf812550c9bbb0631c785b0016a727b38
Reviewed-on: http://gerrit.causes.com/42396
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>