Don't display hook skip message unless hook would run
Previously, we would always display "Skipping <hook name>" when skipping
via the `SKIP` environment variable, even if that hook would never have
run (e.g. due to no applicable files).

This would get annoying when specifying `SKIP=all` as it would display a
large list of hooks that were "skipped"--most of which would never
actually have run.

Change the behavior to only display the message if the hook would have
actually run. In the process, we were able to simplify some boolean
expressions in `Hook#run?` and `HookRunner#run_hooks` as we realized
they didn't need to be as complicated.

Change-Id: I9f503b0d58bd6e3dd5c317f122c8694f3fb05072
Reviewed-on: http://gerrit.causes.com/45273
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>