Remove interactive prompt support from hooks
When we initially added this feature, it wasn't realized that
explicitly trying to reopen the TTY console was a bad idea. Since
there's no guarantee that Overcommit will be run from a TTY context,
having this code in place could cause a hook to hang. It also goes
against the philosophy of how git hooks were intended to be designed
(i.e. they are not supposed to be interactive).

This required a number of changes, namely converting the hook signing
flow to use the CLI `overcommit --sign <hook-type>` instead of being
done interactively.

Change-Id: Ic1176243377aa816039cae45805119e9321bf67e
Reviewed-on: http://gerrit.causes.com/45332
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>