Add .projections.json configuration file
Editor plugins, like @tpope's vim-projectionist, can make use of this
configuration file to enable easy jumping between "alternate" files,
among other features. For example, with this configuration and
vim-projectionist installed, you can now be editing a hook and execute
`:A` to jump to that hook's spec file. This will save some time when
working with this project.

https://github.com/tpope/vim-projectionist

Change-Id: I0500e1e6aad52e1f4da44c4ce674215a2b31f36e
Reviewed-on: http://gerrit.causes.com/45429
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>