Check in repo root for executables
If a project has a local copy of a tool installed, such as jshint via
npm, they may want to configure overcommit to run the locally installed
version. However, overcommit does not currently look in the repo root
when determining if the executable can be run. This commit fixes this by
adding an empty string to the array of paths that are iterated over when
determining if a command is in path.

Change-Id: I38804b3c9766f904c8ebfbb21b8a30827f9facbf
Reviewed-on: http://gerrit.causes.com/45700
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>