Add hook configuration options for flags
In the process of implementing the `command` option, it became evident
that the list of command line flags passed to an executable should also
be configurable.

This is due to the fact that third-party projects change all the time,
and so their flags could be removed/renamed, putting the onus on
Overcommit to implement a fix immediately in order to unbreak hooks.

This is not ideal. Instead, we allow the flags passed to the command to
be customized so that should external tools change users have a path to
immediately unblock themselves rather than wait on Overcommit.

The other benefit here is that it's easy to get a sense of what a hook
is doing by looking at its config, as we now show lots of information
such as the executable, the command executed, and the command line
flags.

Change-Id: I6eef2bf5bb69de2ed0bfd4a7f5c4385da583199f
Reviewed-on: http://gerrit.causes.com/46313
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>