Consider all lines as modified in RunAll context
The `RunAll` context was originally implemented to treat all lines in a
file as having not been modified.

Considering the purpose of this flag is intended to be used for CI runs,
this behavior doesn't make sense because we actually want to scrutinize
the entire repository. However, if the option to ignore lines you didn't
modify is enabled for a hook, this presents a problem since none of the
errors reported in the `RunAll` context will be considered an error.

Change-Id: I9173246e8cce963a271ea8e252192d1ef9f8fabc
Reviewed-on: http://gerrit.causes.com/46484
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>