Don't use puts in GitGuilt hook
Using `puts` violates Overcommit's expectations about when output is
displayed. Capture output to a string and return that instead.

While the intention here is to display output without implying a
warning, unfortunately Overcommit assumes output only comes from hooks
with warnings/errors. For now, we change the hook to report a warning
when displaying any output. We'll have to rethink this system to make
sense for hooks that pass but still want to display output.

Change-Id: I414685af7e5bedf6f6aca7e176de34012f2b521e
Reviewed-on: http://gerrit.causes.com/46964
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>