Replace old array values in configuration instead of append
This was a decision that was made long ago but in hindsight is not wise.

When modifying a configuration value where the value defined in the
default configuration is an array, we would append the new value to the
old array rather than replacing it entirely.

This would result in unexpected behavior when modifying command line
flags for a hook. Instead, we should allow the user to always provide an
explicit list.

Change-Id: I142eca77c8e9eda07610fa499f91e0533a26b33b
Reviewed-on: http://gerrit.causes.com/46967
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>