Add execute_in_background helper
In preparation for the creation of a number of IndexTags hooks for
various hook types, replace the one-off `Process.detach` in `IndexTags`
with the use of an extracted helper `execute_in_background`.

In order to properly test this, a `wait_until` helper was added so we
could check the status of the child process to ensure it terminated
before checking whether it actually executed the specified command.

Change-Id: Ic41ee2b2f4dcb60b0b7cb8b39abf983fd15ac140
Reviewed-on: http://gerrit.causes.com/46990
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>