Ignore hard tabs in Godeps.json
Godep is a utility for managing Go dependencies. It uses the default
JSON marshaller to automatically save this file (should never be
hand-edited). We should not error out or this, since we have no control
over it (and even if we un-tabbed it, it would be re-tabbed next time we
change our dependencies).