Ignore hard tabs in Godeps.json
Godep is a utility for managing Go dependencies. It uses the default
JSON marshaller to automatically save this file (should never be
hand-edited). We should not error out or this, since we have no control
over it (and even if we un-tabbed it, it would be re-tabbed next time we
change our dependencies).

Change-Id: I7612be3cea573acaa659d1653b635cc347a5ac24
Reviewed-on: http://gerrit.causes.com/47264
Reviewed-by: Joe Lencioni <joe.lencioni@brigade.com>
Tested-by: jenkins <jenkins@brigade.com>