Handle custom scss-lint exit statuses
The `scss-lint` executable now returns a special status code when all
files that were specified were filtered by exclusions specified by its
configuration. Previously, this would return an error code that was
ambiguous, but it now returns a custom error code specifically for this
case.

Handling it in Overcommit allows us to save developers the hassle of
duplicating their exclusion configuration from `.scss-lint.yml` to
`.overcommit.yml`.

In the process, we also fixed the hook to always fail unless the status
code indicates lint warnings/errors, since that means the tool was
invoked incorrectly or some other problem is present.