Change problem_on_unmodified_line default to `report`
When attempting to commit a file with hook warnings/errors on lines you
didn't modify, the previous default had Overcommit simply display these
problems as warnings.

However, in some situations this would lead to problems that were
introduced by a commit to still be treated as a warning. For example,
the `Rubocop` `MethodLength` cop reports an error on the method
definition line, but you typically add lines after that line so you may
only get a warning from Overcommit instead of an error.

Change the default to `report` will prevent these sorts of errors from
sneaking in, and results in Overcommit's default `PreCommit`
configuration acting more like its `RunAll` context, where all lines are
considered.

Users will still be able to switch back to `warn` as they desire, even
enabling it wholesale for all pre-commit hooks by setting it in the
`ALL` hook.